refactor: Move TFModelList and TFModelObject functions to conversion package
